+++ 11.一键分析你的上网行为(web页面可视化)/README.MD
@@ -1,0 +1,26 @@
+## Windows Vista, Windows 7,  Windows 8, Windows 10
+历史记录文件位置: `C:\Users\%USERNAME%\AppData\Local\Google\Chrome\User Data\Default\History`
+**注意说明**: `%USERNAME%`为你的用户名
+
+
+
+## Windows XP
+
+历史记录文件位置: `C:\Documents and Settings\%USERNAME%\Local Settings\Application Data\Google\Chrome\User Data\Default\History`
+**注意说明**: `%USERNAME%`为你的用户名
+
+
+
+
+## Mac OS X
+
+历史记录文件位置: `~/Library/Application Support/Google/Chrome/Default/History`
+拷贝历史记录文件到桌面: `cp ~/Library/Application\ Support/Google/Chrome/Default/History ~/Desktop/History`
+**注意说明**: `Application Support`中的空格需要转义，所以改为`Application\ Support`
+
+
+
+## Linux/Unix
+历史记录文件位置:  `~/.config/google-chrome/Default/History`
+拷贝历史记录文件到桌面: `cp ~/.config/google-chrome/Default/History ~/Desktop/History`
+**注意说明**: `如果提示路径不存在, 请自行获取History文件`
