+++ 11.一键分析你的上网行为(web页面可视化)/assets/custom_styles.css
@@ -1,0 +1,11 @@
+/* Removes Undo Button
+–––––––––––––––––––––––––––––––––––––––––––––––––– */
+
+
+._dash-undo-redo {
+	display: none;
+}
+
+html {
+	overflow-y: scroll;
+}
