--- 11.一键分析你的上网行为/README.MD
@@ -1,8 +1,0 @@
-## Linux
-~/.config/google-chrome/Default/History
-
-## MacOS
-~/Library/Application\ Support/Google/Chrome/Default/History
-
-## Windows 7/ Windows 8/ Windows 10
-~\AppData\Local\Google\Chrome\User Data\Default\History
