--- 11.一键分析你的上网行为/assets/custom_styles.css
@@ -1,11 +1,0 @@
-/* Removes Undo Button
-–––––––––––––––––––––––––––––––––––––––––––––––––– */
-
-
-._dash-undo-redo {
-	display: none;
-}
-
-html {
-	overflow-y: scroll;
-}
