--- 11.一键分析你的上网行为/requirement.txt
@@ -1,1 +1,0 @@
-dash == 1.1.0
